# docs/openapi/customer.yaml
openapi: 3.1.0
info:
  title: WatchBox Customer API
  version: 0.1.0
servers:
  - url: http://localhost:4000
...
paths:
...
  /api/v1/users/me:
    get:
      tags: [Customer]
      summary: Get the current customer profile.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current customer profile.
    patch:
      tags: [Customer]
      summary: Update the current customer profile.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated customer profile.
    delete:
      tags: [Customer]
      summary: Delete the current customer account.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Account deleted.
...
  /api/v1/notification-preferences:
    get:
      tags: [Customer]
      summary: Get current customer notification preferences.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current notification preferences.
    patch:
      tags: [Customer]
      summary: Update current customer notification preferences.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationPreferencesRequest"
      responses:
        "200":
          description: Updated notification preferences.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    UpdateProfileRequest:
      type: object
      properties:
        displayName:
          type: string
          minLength: 2
        phone:
          type: string
        country:
          type: string
...
    NotificationPreferencesRequest:
      type: object
      required: [emailAlerts]
      properties:
        emailAlerts:
          type: boolean
...